Refactor a single file to show case how the change looks like
